Moved messaging components to a separate folder and renamed them to be less confusing
