+++ app/renderer/components/messages/Channel.js
