+++ app/renderer/components/messages/Message.js
